+++ app/code/Magento/ImportService/etc/webapi.xml
@@ -8,6 +8,12 @@
 <routes xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Webapi:etc/webapi.xsd">
     <route url="/V1/import" method="POST">
+        <service class="Magento\ImportService\Api\ImportFileInterface" method="importFile"/>
+        <resources>
+            <resource ref="Magento_ImportExport::import" />
+        </resources>
+    </route>
+    <route url="/V1/import/start/:fileId" method="POST">
         <service class="Magento\ImportService\Api\ImportProcessorInterface" method="processImport"/>
         <resources>
             <resource ref="Magento_ImportExport::import" />
